requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-upgrade.txt requirements.in
#
django-djconfig==0.10.0   # via django-spirit
django-haystack==3.0      # via django-spirit
django-infinite-scroll-pagination==1.1.0  # via django-spirit
django-spirit==0.11.1     # via -r requirements.in
django==2.2.16            # via django-haystack, django-spirit
mistune==0.8.4            # via django-spirit
olefile==0.46             # via django-spirit
pillow==7.2.0             # via django-spirit
pytz==2020.1              # via django, django-spirit
sqlparse==0.4.1           # via django
whoosh==2.7.4             # via django-spirit
